- various type adjustments for 64-bit support - removed unused ADT terrain triangle normal parsing

+++ MapViewer/Renderer.hpp
@@ -32,7 +32,7 @@ class Renderer
         {
             CComPtr<ID3D11Buffer> VertexBuffer;
             CComPtr<ID3D11Buffer> IndexBuffer;
-            unsigned int IndexCount;
+            size_t IndexCount;
         };
 
         const HWND m_window;
